pyproject.toml
[tool.poetry]
name = "mattermost-bot"
version = "0.1.0"
description = ""
authors = ["Albrecht Röhm <albrecht@starfish.team>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mattermostdriver = "^7.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
